# server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateProductDto:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.OrderCancelRequest:
    properties:
      status:
        type: string
    type: object
...
  dto.SetProductPriceDto:
    properties:
      price:
        type: number
    type: object
  dto.UpdateProductDto:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
info:
  contact: {}
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Order System
paths:
...
  /api/vendors/orders:
    get:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: itemsPerPage
        type: integer
      - in: query
        name: pageIndex
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all orders of current logged in vendors
      tags:
      - vendor-orders
  /api/vendors/orders/:id:
    put:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: The order has reached its final state
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Move an order to its next status
      tags:
      - vendor-orders
  /api/vendors/orders/:id/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order id
        in: path
        name: id
        required: true
        type: integer
      - description: Cancel order request payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.OrderCancelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Invalid request
        "404":
          description: Order not found (or not belongs to current logged in vendor)
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Cancel an order by logged in vendor
      tags:
      - vendor-orders
  /api/vendors/products:
    get:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: itemsPerPage
        type: integer
      - in: query
        name: pageIndex
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get all products that belongs to the logged in vendor
      tags:
      - vendor-products
    post:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product to be created
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Invalid request
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Create a new product
      tags:
      - vendor-products
  /api/vendors/products/:id:
    put:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update product request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductDto'
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Invalid request
        "403":
          description: Insufficient permission (when try to update a product that
            belongs other vendor)
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Update a product
      tags:
      - vendor-products
  /api/vendors/products/:id/prices:
    get:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "403":
          description: Insufficient permission (when try to get price history of a
            product that belongs other vendor)
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get unit price history of a product
      tags:
      - vendor-products
    post:
      consumes:
      - application/json
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Set product price request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.SetProductPriceDto'
      - description: Product id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "400":
          description: Invalid request
        "403":
          description: Insufficient permission (when try to set price of a product
            that belongs other vendor)
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Set the unit price of a product
      tags:
      - vendor-products
...
swagger: "2.0"